+++ src/OpenSearch/Common/EmptyLogger.php
@@ -24,7 +24,7 @@ namespace OpenSearch\Common;
 use Psr\Log\AbstractLogger;
 use Psr\Log\LoggerInterface;
 
-@trigger_error(__CLASS__ . ' is deprecated in 2.3.2 and will be removed in 3.0.0. Use Psr\Log\NullLogger instead', E_USER_DEPRECATED);
+@trigger_error(EmptyLogger::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0. Use Psr\Log\NullLogger instead', E_USER_DEPRECATED);
 
 /**
  * Class EmptyLogger