+++ src/OpenSearch/ConnectionPool/ConnectionPoolInterface.php
@@ -23,7 +23,7 @@ namespace OpenSearch\ConnectionPool;
 
 use OpenSearch\Connections\ConnectionInterface;
 
-@trigger_error(__CLASS__ . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+@trigger_error(ConnectionPoolInterface::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
 
 /**
  * @deprecated in 2.3.2 and will be removed in 3.0.0.