+++ src/OpenSearch/Connections/ConnectionFactory.php
@@ -24,7 +24,7 @@ namespace OpenSearch\Connections;
 use OpenSearch\Serializers\SerializerInterface;
 use Psr\Log\LoggerInterface;
 
-@trigger_error(__CLASS__ . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+@trigger_error(ConnectionFactory::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
 
 /**
  * @deprecated in 2.3.2 and will be removed in 3.0.0.