+++ src/OpenSearch/Transport.php
@@ -27,7 +27,7 @@ use OpenSearch\ConnectionPool\AbstractConnectionPool;
 use OpenSearch\Connections\ConnectionInterface;
 use Psr\Log\LoggerInterface;
 
-@trigger_error(__CLASS__ . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
+@trigger_error(Transport::class . ' is deprecated in 2.3.2 and will be removed in 3.0.0.', E_USER_DEPRECATED);
 
 /**
  * @deprecated in 2.3.2 and will be removed in 3.0.0.